Speech to Image Generator/requirements.txt
monsterapi
python-dotenv
requests
